Autoresize multiline input height on getting layout
fixes behavior in e.g. an HTML popover when the `<textarea>` renders initially in a `display: none` ancestor and therefore has no dimensions when the initial `useEffect` runs. with this, when it becomes visible, the ResizeObserver fires and we can recalculate the height. it should also fire if there is for example a whileInView transition that takes it from collapsed to visible, or any other scenario where it initially renders without dimensions.