📖 Add dropdown Fixed Header story
in debugging anchor positioning with a fixed or absolute or relative positioned parent, found that the resulting offset parent needs room in both dimensions for a position fallback to be preferred over the default. in other words, if there is more room to the right than to the left, but there isn’t enough room vertically, the left fallback position won’t be chosen even though it would be better for one of the two dimensions, whereas if there is enough room vertically, it will be chosen.